what we did in this commit:
home page fetchs last 4th invites -m we faced two problem here: -m 1.home page will be rendered either statically or by ssr. in both cases if the user is not yet logged in, we will get jwt error. -m to fix this problem we use hasura admin-secret. it is just used on server side for rendering and will not be delivered to client. -m the other problem was to share the jwt token between client and server. we used cookies for that